refactor: move database connector interface
