--- packages/backend-common/src/database/connector.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright 2021 Spotify AB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { Config } from '@backstage/config';
-import { Knex } from 'knex';
-
-export interface DatabaseConnector {
-  createClient(dbConfig: Config, overrides?: Partial<Knex.Config>): Knex;
-  createNameOverride(name: string): Partial<Knex.Config>;
-  parseConnectionString(
-    connectionString: string,
-    client?: string,
-  ): Knex.StaticConnectionConfig;
-  ensureDatabaseExists?(
-    dbConfig: Config,
-    ...databases: Array<string>
-  ): Promise<void>;
-}
