+++ packages/backend-common/src/database/types.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import { Config } from '@backstage/config';
 import { Knex } from 'knex';
 
 /**
@@ -27,4 +28,38 @@ export interface PluginDatabaseManager {
    * stores so that plugins are discouraged from database integration.
    */
   getClient(): Promise<Knex>;
+}
+
+/**
+ * DatabaseConnector manages an underlying Knex database driver.
+ */
+export interface DatabaseConnector {
+  /**
+   * createClient provides an instance of a knex database connector.
+   */
+  createClient(dbConfig: Config, overrides?: Partial<Knex.Config>): Knex;
+  /**
+   * createNameOverride provides a partial knex config sufficient to override a
+   * database name.
+   */
+  createNameOverride(name: string): Partial<Knex.Config>;
+  /**
+   * parseConnectionString produces a knex connection config object representing
+   * a database connection string.
+   */
+  parseConnectionString(
+    connectionString: string,
+    client?: string,
+  ): Knex.StaticConnectionConfig;
+  /**
+   * ensureDatabaseExists performs a side-effect to ensure database names passed in are
+   * present.
+   *
+   * Calling this function on databases which already exist should do nothing.
+   * Missing databases should be created if needed.
+   */
+  ensureDatabaseExists?(
+    dbConfig: Config,
+    ...databases: Array<string>
+  ): Promise<void>;
 }
